+++ packages/core/src/types.ts
@@ -20,7 +20,7 @@ export type ThemeName = string;
 
 export type ClassName = string;
 
-export type ClassNameGenerator<N extends object, P extends object | string> = (
+export type ClassNameTransformer<N extends object, P extends object | string> = (
   ...styles: (undefined | false | ClassName | N | P)[]
 ) => ClassName;