+++ packages/core/tests/Aesthetic.test.tsx
@@ -42,6 +42,7 @@ describe('Aesthetic', () => {
       });
 
       expect(instance.options).toEqual({
+        cxPropName: 'cx',
         extendable: false,
         passThemeProp: false,
         pure: true,
@@ -477,37 +478,40 @@ describe('Aesthetic', () => {
   describe('transformStyles()', () => {
     it('errors for invalid value', () => {
       expect(() => {
-        instance.transformStyles([true]);
+        instance.transformStyles(
+          // @ts-ignore Allow invalid type
+          [123],
+        );
       }).toThrowErrorMatchingSnapshot();
     });
 
     it('combines strings into a class name', () => {
-      expect(instance.transformStyles('foo', 'bar')).toBe('foo bar');
+      expect(instance.transformStyles(['foo', 'bar'])).toBe('foo bar');
     });
 
     it('calls `transformToClassName` method', () => {
       // @ts-ignore Allow access
       const spy = jest.spyOn(instance, 'transformToClassName');
 
-      instance.transformStyles({ color: 'red' }, { display: 'block' });
+      instance.transformStyles([{ color: 'red' }, { display: 'block' }]);
 
       expect(spy).toHaveBeenCalledWith([{ color: 'red' }, { display: 'block' }]);
     });
 
     it('ignores falsey values', () => {
-      expect(instance.transformStyles(null, false, 0, '', undefined)).toBe('');
+      expect(instance.transformStyles([null, false, 0, '', undefined])).toBe('');
     });
 
     it('strips period prefix', () => {
-      expect(instance.transformStyles('.foo', 'bar .qux')).toBe('foo bar qux');
+      expect(instance.transformStyles(['.foo', 'bar .qux'])).toBe('foo bar qux');
     });
 
     it('handles expression values', () => {
-      expect(instance.transformStyles('foo', true && 'bar', 5 > 10 && 'baz')).toBe('foo bar');
+      expect(instance.transformStyles(['foo', true && 'bar', 5 > 10 && 'baz'])).toBe('foo bar');
     });
 
     it('joins strings', () => {
-      expect(instance.transformStyles('foo', '123', 'bar')).toBe('foo 123 bar');
+      expect(instance.transformStyles(['foo', '123', 'bar'])).toBe('foo 123 bar');
     });
   });
 
@@ -520,7 +524,7 @@ describe('Aesthetic', () => {
 
       const styleSheet = adapter.createStyleSheet('foo');
 
-      expect(adapter.transformStyles(styleSheet.button)).toMatchSnapshot();
+      expect(adapter.transformStyles([styleSheet.button])).toMatchSnapshot();
     });
 
     it('supports Aphrodite', () => {
@@ -531,7 +535,7 @@ describe('Aesthetic', () => {
 
       const styleSheet = adapter.createStyleSheet('foo');
 
-      expect(adapter.transformStyles(styleSheet.button)).toMatchSnapshot();
+      expect(adapter.transformStyles([styleSheet.button])).toMatchSnapshot();
     });
 
     it('supports CSS modules', () => {
@@ -542,7 +546,7 @@ describe('Aesthetic', () => {
 
       const styleSheet = adapter.createStyleSheet('foo');
 
-      expect(adapter.transformStyles(styleSheet.button)).toMatchSnapshot();
+      expect(adapter.transformStyles([styleSheet.button])).toMatchSnapshot();
     });
 
     it('supports Fela', () => {
@@ -557,7 +561,7 @@ describe('Aesthetic', () => {
 
       const styleSheet = adapter.createStyleSheet('foo');
 
-      expect(adapter.transformStyles(styleSheet.button)).toMatchSnapshot();
+      expect(adapter.transformStyles([styleSheet.button])).toMatchSnapshot();
     });
 
     it('supports JSS', () => {
@@ -571,7 +575,7 @@ describe('Aesthetic', () => {
 
       const styleSheet = adapter.createStyleSheet('foo');
 
-      expect(adapter.transformStyles(styleSheet.button)).toMatchSnapshot();
+      expect(adapter.transformStyles([styleSheet.button])).toMatchSnapshot();
     });
 
     it('supports TypeStyle', () => {
@@ -582,7 +586,7 @@ describe('Aesthetic', () => {
 
       const styleSheet = adapter.createStyleSheet('foo');
 
-      expect(adapter.transformStyles(styleSheet.button)).toMatchSnapshot();
+      expect(adapter.transformStyles([styleSheet.button])).toMatchSnapshot();
     });
   });
 });