+++ packages/react/CHANGELOG.md
@@ -7,3 +7,10 @@
 #### 💥 Breaking
 
 - Updated `react` requirement to v16.6.
+- **[TS]** Renamed the `WithStylesProps` interface to `WithStylesWrappedProps`.
+- **[TS]** Renamed the `WithThemeProps` interface to `WithThemeWrappedProps`.
+
+#### 🚀 Updates
+
+- Added a `cxPropName` option to `withStyles`.
+- Updated `withStyles` HOC to receive the CSS transformer function as a `cx` prop.